wallet: fix `gosec` linter errors on int32 conversion
We now properly fix the `int32 <-> uint32` conversion in the new code -
for the old code, it's safe to ignore cause they will be removed
eventually.

+++ wallet/account_manager_benchmark_test.go
@@ -466,8 +466,8 @@ func BenchmarkGetBalanceAPI(b *testing.B) {
 
 			for b.Loop() {
 				_, err := bw.Balance(
-					b.Context(), confirmations, scopes[0],
-					accountName,
+					b.Context(), uint32(confirmations),
+					scopes[0], accountName,
 				)
 				require.NoError(b, err)
 			}
